Introduce css inside main html file
This commit will introduce css styling inside of an html file. I added styles to individual parts of the body. I  added clickable buttons that take you to certain recipes on the relative webpage. Centered everything on the webpage. Added hover and active features to the buttons as well as tranistions when hovering over the buttons. Included a background picture to the main webpage.